kbuilder/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "kbdr-kbuilder"
description = "KBDr Kernel Builder"
readme = "README.md"
requires-python = ">=3.11"
version = "v0.3.5"
dependencies = [
    "jsonschema",
    "google-auth",
    "google-cloud-storage",
    "aio_pika",
    "aiofiles"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["KBDr.*"]  # ["*"] by default
